requirements.txt
firebase-admin
openai
groq
fastapi
uvicorn[standard]
httpx
pypdf
python-dotenv
